Update contact form to use reCAPTCHA v3 and adjust styling
Replaced the reCAPTCHA v2 widget with reCAPTCHA v3 in the contact form for improved user experience. Updated the required score in settings.py to lower the threshold suitable for v3. Removed unnecessary CSS styling for the reCAPTCHA widget, as v3 is invisible. This enhances the form's functionality and aligns with the latest reCAPTCHA standards.